+++ test/fixtures/comma-align-with-max.sql
@@ -1,6 +1,6 @@
 create table "comma-align-with-max" (
  "text-field" character varying(15),
- "int-field" integer(2),
+ "int-field" integer,
  "numerico-el-1" numeric(9,6),
  primary key ("text-field", "int-field")
 );
